Terraform_Code/requirements.txt
python-terraform
pyyaml
subprocess.run